+++ AppUI/UIComponents/Composed/CardSearchPreviewViewController.py
@@ -2,7 +2,7 @@
 from typing import List
 
 from PyQt5.QtWidgets import (QLabel, QPushButton, QSizePolicy, QVBoxLayout,
-                             QWidget)
+                             QWidget, QHBoxLayout)
 
 from AppCore.Config import ConfigurationProvider
 from AppCore.Models import CardType, TradingCard
@@ -25,39 +25,48 @@ class CardSearchPreviewViewController(QWidget):
         super().__init__()
 
         layout = QVBoxLayout()
+        layout.setContentsMargins(0, 0, 0, 0)
         self.setLayout(layout)
         self._observation_tower = observation_tower
         self._card_image_source_provider = card_image_source_provider
         # https://stackoverflow.com/a/19011496
-        preview_view = ImagePreviewViewController(observation_tower, 
+        staging_view = ImagePreviewViewController(observation_tower, 
                                                   configuration_provider, 
                                                   asset_provider)
-        preview_view.setMinimumHeight(300)
-        
-        self.staging_view = preview_view
-        preview_view.setSizePolicy(QSizePolicy.Policy.Minimum, QSizePolicy.Policy.Fixed)
+        staging_view.setMinimumHeight(300)
+        staging_view.setSizePolicy(QSizePolicy.Policy.Minimum, QSizePolicy.Policy.Fixed)
         # lbl2.setMinimumHeight(300);
-        layout.addWidget(preview_view)
+        self.staging_view = staging_view
+        layout.addWidget(staging_view)
 
-        retry_button = QPushButton()
-        retry_button.setText("Redownload")
-        # retry_button.setEnabled(False)
-        retry_button.clicked.connect(self.tapped_retry_button)
-        self.retry_button = retry_button
-        layout.addWidget(retry_button)
+        buttons_layout = QHBoxLayout()
+        buttons_widget = QWidget()
+        # buttons_widget.setStyleSheet('background-color: green')
+        buttons_widget.setLayout(buttons_layout)
+        layout.addWidget(buttons_widget)
+
 
         flip_button = QPushButton()
         flip_button.setText("Flip (Ctrl+F)")
         flip_button.setEnabled(False)
         flip_button.clicked.connect(self.tapped_flip_button)
         self.flip_button = flip_button
-        layout.addWidget(flip_button)
+        buttons_layout.addWidget(flip_button)
+        
+        retry_button = QPushButton()
+        retry_button.setText("Redownload")
+        retry_button.setEnabled(False)
+        retry_button.clicked.connect(self.tapped_retry_button)
+        self.retry_button = retry_button
+        buttons_layout.addWidget(retry_button)
+        
         
         search_table_view = SearchTableView(observation_tower, 
                                             card_type_list)
+        # search_table_view.setStyleSheet('background-color: yellow')
         search_table_view.delegate = self
         self.search_table_view = search_table_view
-        layout.addWidget(search_table_view)
+        layout.addWidget(search_table_view, 2)
         
         image_source_label = QLabel()
         image_source_label.setOpenExternalLinks(True)
@@ -98,6 +107,7 @@ class CardSearchPreviewViewController(QWidget):
         self._current_image_path = local_resource.image_preview_path
         self.staging_view.set_image(local_resource)
         self.flip_button.setEnabled(is_flippable)
+        self.retry_button.setEnabled(local_resource.remote_image_url is not None)
         
 
     def update_list(self, result_list: List[TradingCard]):
@@ -114,14 +124,6 @@ class CardSearchPreviewViewController(QWidget):
         self.image_source_label.setText(f'Image source: <a href="{url}">{url}</a>')
         
     def handle_observation_tower_event(self, event: TransmissionProtocol):
-        if type(event) == LocalResourceEvent:
-            if self._current_image_path is not None:
-                pass
-        elif type(event) == ConfigurationUpdatedEvent:
+        if type(event) == ConfigurationUpdatedEvent:
             self._load_image_source_label()
-                # if self._current_image_path == event.local_resource.image_preview_path:
-                #     if event.event_type == LocalResourceEvent.EventType.STARTED:
-                #         self.retry_button.setEnabled(False)
-                #     elif event.event_type == LocalResourceEvent.EventType.FINISHED:
-                #         self.retry_button.setEnabled(True)